Moves PHPUnit compatability code to our test namespace
It's not something we want to support
