+++ library/Mockery/Adapter/Phpunit/MockeryTestCase.php
@@ -32,22 +32,4 @@ abstract class MockeryTestCase extends \PHPUnit\Framework\TestCase
     protected function mockeryTestTearDown()
     {
     }
-
-    public function expectExceptionMessageRegEx($regularExpression)
-    {
-        if (method_exists(get_parent_class(), 'expectExceptionMessageRegExp')) {
-            return parent::expectExceptionMessageRegExp($regularExpression);
-        }
-
-        return $this->expectExceptionMessageMatches($regularExpression);
-    }
-
-    public static function assertMatchesRegEx($pattern, $string, $message = '')
-    {
-        if (method_exists(get_parent_class(), 'assertMatchesRegularExpression')) {
-            parent::assertMatchesRegularExpression($pattern, $string, $message);
-        }
-
-        self::assertRegExp($pattern, $string, $message);
-    }
 }
