+++ tests/Mockery/RegExpCompatability.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace test\Mockery;
+
+trait RegExpCompatability 
+{
+    public function expectExceptionMessageRegEx($regularExpression)
+    {
+        if (method_exists(get_parent_class(), 'expectExceptionMessageRegExp')) {
+            return parent::expectExceptionMessageRegExp($regularExpression);
+        }
+
+        return $this->expectExceptionMessageMatches($regularExpression);
+    }
+
+    public static function assertMatchesRegEx($pattern, $string, $message = '')
+    {
+        if (method_exists(get_parent_class(), 'assertMatchesRegularExpression')) {
+            parent::assertMatchesRegularExpression($pattern, $string, $message);
+        }
+
+        self::assertRegExp($pattern, $string, $message);
+    }
+}
